Move GetAccountDecoder to prevent import cycle and replace calls to it with one call in WithAccountDecoder
